integração front e back - pagina de posts e pagina de admin

+++ Front-End/app/_components/post/PostList.tsx
@@ -1,12 +1,12 @@
 "use client";
 
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { format } from "date-fns";
 import { ptBR } from "date-fns/locale";
-import { Edit2, Trash2 } from "lucide-react";
+import { Ban, Edit2, Trash2 } from "lucide-react";
 import Link from "next/link";
-import PostListLoading from "../loadings/PostListLoading";
+import { useRouter, useSearchParams } from "next/navigation";
 import QueryError from "../errors/QueryError";
 import { useToast } from "../ui/use-toast";
 import { IPost } from "@/app/api/_services/modules/post/entities/Post";
@@ -14,21 +14,44 @@ import { Column, DataTable } from "../shared/DataTable";
 import { Button } from "../ui/button";
 import { PostDialog } from "./dialogs/PostDialog";
 import { Badge } from "../ui/badge";
+import {
+   AlertDialog,
+   AlertDialogAction,
+   AlertDialogCancel,
+   AlertDialogContent,
+   AlertDialogDescription,
+   AlertDialogFooter,
+   AlertDialogHeader,
+   AlertDialogTitle,
+   AlertDialogTrigger,
+} from "../ui/alert-dialog";
+import { ITENS_PER_PAGE_TABLE } from "@/utils/constantes/constants";
+import { CategoryBadge } from "../category/CategoryBadge";
+import { LoadingDataTable } from "../loadings/PostListLoading";
 
-const PAGE_SIZE = 10;
+interface IPostList {
+   currentPage: number;
+   setCurrentPage: (number: number) => void;
+}
 
-export function PostList() {
-   const [page, setPage] = useState(1);
+export function PostList({ currentPage, setCurrentPage }: IPostList) {
    const { toast } = useToast();
    const queryClient = useQueryClient();
 
    const { data, isLoading, error, refetch } = useQuery({
-      queryKey: ["posts"],
+      queryKey: ["posts", currentPage],
       queryFn: async () => {
-         const response = await fetch("/api/posts");
+         await new Promise((resolve) => setTimeout(resolve, 0.1 * 1000));
+         const params = new URLSearchParams({
+            page: currentPage.toString(),
+            limit: ITENS_PER_PAGE_TABLE.toString(),
+         });
+
+         const response = await fetch(`/api/posts/page?${params}`);
 
          if (!response.ok) {
-            throw new Error("Erro ao buscar categorias");
+            const error = await response.json();
+            throw new Error(error?.message || "Erro Desconhecido");
          }
 
          return response.json();
@@ -37,10 +60,13 @@ export function PostList() {
 
    const { mutate: deletePost } = useMutation({
       mutationFn: async (id: string) => {
-         const response = await fetch(`/api/posts?id=${id}`);
+         const response = await fetch(`/api/posts/${id}`, {
+            method: "DELETE",
+         });
 
          if (!response.ok) {
-            throw new Error("Erro ao buscar categorias");
+            const error = await response.json();
+            throw new Error(error?.message || "Erro Desconhecido");
          }
          return response.json();
       },
@@ -49,35 +75,35 @@ export function PostList() {
             title: "Post excluído com sucesso!",
             description: "O post foi removido do sistema.",
          });
-         queryClient.invalidateQueries({ queryKey: ["posts"] });
+         refetch();
       },
-      onError: () => {
+      onError: (error) => {
          toast({
             title: "Erro ao excluir post",
-            description: "Ocorreu um erro ao tentar excluir o post.",
+            description:
+               "Ocorreu um erro ao tentar excluir o post: " + error.message,
             variant: "destructive",
          });
       },
    });
 
    const handleDelete = async (id: string) => {
-      try {
-         await deletePost(id);
-         refetch();
-      } catch (error) {
-         toast({
-            title: "Erro ao excluir",
-            description: "Ocorreu um erro ao tentar excluir o post.",
-            variant: "destructive",
-         });
-      }
+      deletePost(id);
    };
 
-   if (isLoading) return <PostListLoading />;
+   const handlePageChange = (newPage: number) => {
+      setCurrentPage(newPage);
+   };
+
+   if (isLoading) return <LoadingDataTable />;
 
    if (error) return <QueryError onRetry={() => refetch()} />;
 
    const columns: Column<IPost>[] = [
+      {
+         header: "ID",
+         accessorKey: (post: IPost) => post.id,
+      },
       {
          header: "Título",
          accessorKey: (post: IPost) => post.title,
@@ -86,17 +112,10 @@ export function PostList() {
          header: "Categoria",
          accessorKey: (post: IPost) => {
             return (
-               <div className="flex gap-x-2 text-primary">
+               <div className="flex flex-wrap w-56 gap-2 text-primary">
                   {post.categories?.length !== 0 ? (
                      post?.categories?.map((category) => {
-                        return (
-                           <Badge
-                              style={{ backgroundColor: category.color }}
-                              className={`text-primary`}
-                           >
-                              {category.name}
-                           </Badge>
-                        );
+                        return <CategoryBadge category={category} />;
                      })
                   ) : (
                      <span>Sem Categoria</span>
@@ -121,36 +140,55 @@ export function PostList() {
          header: "Ações",
          accessorKey: (post: IPost) => (
             <div className="flex items-center gap-2">
-               <PostDialog post={post} mode="edit">
+               <PostDialog currentPage={currentPage} mode="edit" post={post}>
                   <Button variant="ghost" size="icon">
                      <Edit2 className="h-4 w-4" />
-                     <span className="sr-only">Editar post</span>
                   </Button>
                </PostDialog>
-               <Button
-                  variant="ghost"
-                  size="icon"
-                  className="text-destructive"
-                  onClick={() => handleDelete(post.id)}
-               >
-                  <Trash2 className="h-4 w-4" />
-                  <span className="sr-only">Excluir post</span>
-               </Button>
+               <AlertDialog>
+                  <AlertDialogTrigger asChild>
+                     <Button variant="ghost" size="icon">
+                        <Trash2 className="h-4 w-4 text-destructive" />
+                     </Button>
+                  </AlertDialogTrigger>
+                  <AlertDialogContent>
+                     <AlertDialogHeader>
+                        <AlertDialogTitle>
+                           Tem certeza que deseja excluir este post?
+                        </AlertDialogTitle>
+                        <AlertDialogDescription>
+                           Esta ação não pode ser desfeita. Isso excluirá
+                           permanentemente o post e removerá os dados do
+                           servidor.
+                        </AlertDialogDescription>
+                     </AlertDialogHeader>
+                     <AlertDialogFooter>
+                        <AlertDialogCancel>Cancelar</AlertDialogCancel>
+                        <AlertDialogAction
+                           onClick={() => handleDelete(post.id)}
+                        >
+                           Continuar
+                        </AlertDialogAction>
+                     </AlertDialogFooter>
+                  </AlertDialogContent>
+               </AlertDialog>
             </div>
          ),
       },
    ];
 
    return (
-      <DataTable<IPost>
-         data={data || []}
-         columns={columns}
-         pagination={{
-            page,
-            pageSize: PAGE_SIZE,
-            total: data?.length || 0,
-         }}
-         onPageChange={setPage}
-      />
+      <div className="space-y-4">
+         <DataTable<IPost>
+            data={data?.posts || []}
+            columns={columns}
+            pagination={{
+               page: currentPage,
+               pageSize: ITENS_PER_PAGE_TABLE,
+               total: data?.total || 0,
+            }}
+            onPageChange={handlePageChange}
+         />
+      </div>
    );
 }
